MIPS Assembler - Command Line Project
## Description

The project is a C++ project created inside xCode.

This will load a text file that contains mips instructions and convert
it into machine code.

Please see the ‘MIPS_Green_Sheet.pdf’ for what instructions it can
convert. The only type of instructions that it will convert are:

- R-Type
- I-Type

The format for an input file is attached in ‘test_1.s’ and will output
in a file called: ‘output.txt’

The original project description is in Project1.docx

## Running Project

- Load this project into any C++ complier.

- Build the project

- Find the executable

- ./myAssembler <input.txt>

- Find ‘Output.txt’